// src/models/Languages/Languages.ts
export enum LanguageCode {
    UNKNOWN = "_unknown",
    BENGALI = "bd",
    BULGARIAN = "bg",
    BRAZILIAN = "br",
    CHINEESE = "cn",
    CZECH = "cz",
    GERMAN = "de",
    DANISH = "dk",
    ENGLISH = "gb",
    SPANISH = "es",
    FINNISH = "fi",
    FRENCH = "fr",
    WELSH = "gb",
    GREEK = "gr",
    CHINEESE_HONGKONG = "hk",
    HUNGARIAN = "hu",
    INDONESIAN = "id",
    ISRELI = "il",
    INDIAN = "in",
    IRAN = "ir",
    ITALIAN = "it",
    JAPANESE = "jp",
    KOREAN = "kr",
    LITHUANIAN = "lt",
    MONGOLIAN = "mn",
    MEXIAN = "mx",
    MALAY = "my",
    DUTCH = "nl",
    NORWEGIAN = "no",
    PHILIPPINE = "ph",
    POLISH = "pl",
    PORTUGUESE = "pt",
    ROMANIAN = "ro",
    RUSSIAN = "ru",
    SANSKRIT = "sa", // in the year of our lord >2000 A.D
    SAMI = "si",
    THAI = "th",
    TURKISH = "tr",
    UKRAINIAN = "ua",
    VIETNAMESE = "vn",
}